fixed the edge runtime config
