+++ package.json
@@ -43,7 +43,7 @@
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@react-email/components": "^0.0.21",
+    "@react-email/components": "^0.0.14",
     "@react-email/render": "^0.0.16",
     "@react-pdf/renderer": "^3.4.4",
     "@sindresorhus/slugify": "^2.2.1",
