+++ src/app/sitemap.ts
@@ -2,16 +2,21 @@ import type { MetadataRoute } from "next";
 
 export const baseUrl = "https://jani-ai.com";
 
-// export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-//   const blogs = getBlogPosts().map((post) => ({
-//     url: `${baseUrl}/blog/${post.slug}`,
-//     lastModified: post.metadata.publishedAt,
-//   }));
+import { headers } from "next/headers";
 
-//   const routes = ["", "/blog"].map((route) => ({
-//     url: `${baseUrl}${route}`,
-//     lastModified: new Date().toISOString().split("T")[0],
-//   }));
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+  const headersList = headers();
+  let domain = headersList.get("host") as string;
 
-//   return [...routes, ...blogs];
-// }
+  if (domain === "localhost:3000" || domain.endsWith(".vercel.app")) {
+    // for local development and preview URLs
+    domain = "localhost:3000";
+  }
+
+  return [
+    {
+      url: `https://${domain}`,
+      lastModified: new Date(),
+    },
+  ];
+}
