+++ src/utils/lib/auth/auth.config.ts
@@ -6,8 +6,7 @@ import { db } from "~/server/db";
 import { JWT } from "next-auth/jwt";
 import { User } from "next-auth";
 import { AdapterUser } from "next-auth/adapters";
-import { resend } from "~/emails";
-import WelcomeEmail from "~/emails/welcome-email";
+
 
 export default {
   providers: [
@@ -101,53 +100,7 @@ export default {
       return session;
     },
   },
-  events: {
-    async signIn(message) {
-      if (message.isNewUser) {
-        const email = message.user.email as string;
-        const user = await db.user.findUnique({
-          where: { email },
-          select: {
-            id: true,
-            name: true,
-            email: true,
-            image: true,
-            createdAt: true,
-          },
-        });
-        if (!user) {
-          return;
-        }
-        // only send the welcome email if the user was created in the last 10s
-        // (this is a workaround because the `isNewUser` flag is triggered when a user does `dangerousEmailAccountLinking`)
-        if (
-          user.createdAt &&
-          new Date(user.createdAt).getTime() > Date.now() - 10000
-        ) {
-          resend.emails.send({
-            subject: "Welcome to Jani AI",
-            to: email,
-            react: WelcomeEmail({
-              email,
-              name: user?.name || null,
-              marketing: true,
-            }),
-            from: "brighton.mboya.io@gmail.com",
-          });
+  
 
-          // await Promise.allSettled([
-          //   sendEmail({
-          //     subject: "Welcome to Jani AI!",
-          //     email,
-          //     react: WelcomeEmail({
-          //       email,
-          //       name: user.name || null,
-          //     }),
-          //     marketing: true,
-          //   }),
-          // ]);
-        }
-      }
-    },
-  },
+  
 } satisfies NextAuthConfig;
